up: InputContainer 분리 및 그에 따른 Page test 수정
